requirements/production.txt
Django
djangorestframework
uritemplate
dj_database_url
mysqlclient
